Improved implementation of ticket #601 - Fixed an issue with the "print" command - Resolved a bunch of compiler warnings
git-svn-id: https://svn.code.sf.net/p/numere/source/branches/ticket_601@981 d50cd26f-24d5-421e-9e7d-30b8b4216d91

+++ gui/compositions/tableviewer.cpp
@@ -972,7 +972,7 @@ mu::value_type TableViewer::CellToCmplx(int row, int col)
 {
     if (GetTable()->CanGetValueAs(row, col, wxGRID_VALUE_FLOAT))
         return GetTable()->GetValueAsDouble(row, col);
-    else if (row >= nFirstNumRow && GetCellValue(row, col)[0] != '"' && isNumerical(GetCellValue(row, col).ToStdString()))
+    else if (row >= (int)nFirstNumRow && GetCellValue(row, col)[0] != '"' && isNumerical(GetCellValue(row, col).ToStdString()))
         return StrToCmplx(GetCellValue(row, col).ToStdString());
 
     return NAN;
